--- a/app/helpers/activity_helper.rb
+++ b/app/helpers/activity_helper.rb
@@ -13,7 +13,7 @@ module ActivityHelper
 
   def activity_back_path(current_user:, activity:)
     if activity.programme? && current_user.service_owner?
-      return organisation_activity_path(activity.parent_activity.organisation, activity.parent_activity)
+      return organisation_activity_path(activity.parent.organisation, activity.parent)
     end
 
     organisation_path(current_user.organisation)