--- a/app/services/find_project_activities.rb
+++ b/app/services/find_project_activities.rb
@@ -10,7 +10,7 @@ class FindProjectActivities
 
   def call
     projects = policy_scope(Activity.project, policy_scope_class: ProjectPolicy::Scope)
-      .includes(:organisation)
+      .includes(:organisation, :parent)
       .order("created_at ASC")
     projects = if organisation.service_owner
       projects.all