--- a/app/views/staff/shared/activities/_projects.html.haml
+++ b/app/views/staff/shared/activities/_projects.html.haml
@@ -3,5 +3,5 @@
     %h2.govuk-heading-m
       = t("page_content.activity.projects")
     - if policy(:project).create?
-      = button_to t("page_content.organisation.button.create_project"), organisation_fund_programme_projects_path(current_user.organisation, activity.parent_activity, activity), class: "govuk-button"
+      = button_to t("page_content.organisation.button.create_project"), organisation_fund_programme_projects_path(current_user.organisation, activity.parent, activity), class: "govuk-button"
     = render partial: '/staff/shared/projects/table', locals: { projects: activities }