--- a/config/environments/test.rb
+++ b/config/environments/test.rb
@@ -59,6 +59,8 @@ Rails.application.configure do
     Bullet.enable = true
     Bullet.bullet_logger = true
     Bullet.raise = true # raise an error if n+1 query occurs
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: "User", association: :organisation
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: "Activity", association: :organisation
     Bullet.add_whitelist type: :unused_eager_loading, class_name: "Transaction", association: :provider
     Bullet.add_whitelist type: :unused_eager_loading, class_name: "Transaction", association: :receiver
   end