--- a/spec/features/staff/users_can_create_a_budget_spec.rb
+++ b/spec/features/staff/users_can_create_a_budget_spec.rb
@@ -44,7 +44,7 @@ RSpec.describe "Users can create a budget" do
         programme_activity = create(:programme_activity, parent: fund_activity, organisation: user.organisation)
 
         visit organisation_path(user.organisation)
-        click_on(programme_activity.parent_activity.title)
+        click_on(programme_activity.parent.title)
         click_on I18n.t("tabs.activity.details")
         click_on(programme_activity.title)
 
@@ -61,7 +61,7 @@ RSpec.describe "Users can create a budget" do
 
         visit organisation_path(user.organisation)
 
-        click_on(programme_activity.parent_activity.title)
+        click_on(programme_activity.parent.title)
         click_on I18n.t("tabs.activity.details")
         click_on(programme_activity.title)