--- a/spec/features/staff/users_can_create_a_transaction_spec.rb
+++ b/spec/features/staff/users_can_create_a_transaction_spec.rb
@@ -257,7 +257,7 @@ RSpec.feature "Users can create a transaction" do
     scenario "they cannot create transactions on a programme" do
       fund_activity = create(:fund_activity, organisation: user.organisation)
       programme_activity = create(:programme_activity,
-        activity: fund_activity,
+        parent: fund_activity,
         organisation: user.organisation,
         extending_organisation: user.organisation)