--- a/spec/features/staff/users_can_view_budgets_on_an_activity_page_spec.rb
+++ b/spec/features/staff/users_can_view_budgets_on_an_activity_page_spec.rb
@@ -37,7 +37,7 @@ RSpec.feature "Users can view budgets on an activity page" do
     context "when the activity is programme level" do
       scenario "budget information is shown on the page" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
-        programme_activity = create(:programme_activity, activity: fund_activity, organisation: user.organisation)
+        programme_activity = create(:programme_activity, parent: fund_activity, organisation: user.organisation)
 
         budget = create(:budget, parent_activity: programme_activity)
         budget_presenter = BudgetPresenter.new(budget)
@@ -55,8 +55,8 @@ RSpec.feature "Users can view budgets on an activity page" do
     context "when the activity is project level" do
       scenario "budget information is shown on the page" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
-        programme_activity = create(:programme_activity, activity: fund_activity, organisation: user.organisation)
-        project_activity = create(:project_activity, activity: programme_activity, organisation: user.organisation)
+        programme_activity = create(:programme_activity, parent: fund_activity, organisation: user.organisation)
+        project_activity = create(:project_activity, parent: programme_activity, organisation: user.organisation)
 
         budget = create(:budget, parent_activity: project_activity)
         budget_presenter = BudgetPresenter.new(budget)
@@ -74,8 +74,8 @@ RSpec.feature "Users can view budgets on an activity page" do
 
       scenario "a BEIS user cannot edit/create a budget" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
-        programme_activity = create(:programme_activity, activity: fund_activity, organisation: user.organisation)
-        project_activity = create(:project_activity, activity: programme_activity, organisation: user.organisation)
+        programme_activity = create(:programme_activity, parent: fund_activity, organisation: user.organisation)
+        project_activity = create(:project_activity, parent: programme_activity, organisation: user.organisation)
 
         budget = create(:budget, parent_activity: project_activity)
 
@@ -130,7 +130,7 @@ RSpec.feature "Users can view budgets on an activity page" do
     context "when the activity is project level" do
       scenario "budget information is shown on the page" do
         programme_activity = create(:programme_activity, extending_organisation: user.organisation, organisation: user.organisation)
-        project_activity = create(:project_activity, activity: programme_activity, organisation: user.organisation)
+        project_activity = create(:project_activity, parent: programme_activity, organisation: user.organisation)
 
         budget = create(:budget, parent_activity: project_activity)
         budget_presenter = BudgetPresenter.new(budget)
@@ -146,7 +146,7 @@ RSpec.feature "Users can view budgets on an activity page" do
 
       scenario "a delivery partner can edit/create a budget" do
         programme_activity = create(:programme_activity, extending_organisation: user.organisation, organisation: user.organisation)
-        project_activity = create(:project_activity, activity: programme_activity, organisation: user.organisation)
+        project_activity = create(:project_activity, parent: programme_activity, organisation: user.organisation)
 
         budget = create(:budget, parent_activity: project_activity)