… CHANGELOG.md
… Gemfile.lock
… spec/features/staff/users_can_view_programme_level_activity_spec.rb
… spec/features/staff/users_can_view_transactions_on_an_activity_page_spec.rb
… spec/helpers/activity_helper_spec.rb
… spec/models/activity_spec.rb
… spec/presenters/activity_xml_presenter_spec.rb
… spec/services/create_programme_activity_spec.rb
… spec/services/create_project_activity_spec.rb
… spec/services/create_third_party_project_activity_spec.rb
… spec/services/find_programme_activities_spec.rb
… spec/services/find_project_activities_spec.rb
… spec/services/find_third_party_project_activities_spec.rb
… spec/services/ingest_iati_activities_spec.rb
… spec/support/bullet.rb