## Changes in this PR

- refactor the parent relationship of activities so that instead of `activity.activity_id` we have `activity.parent_id`. This follows a found convention for Rails gems that deal trees and enables the installation of the `acts_as_tree` with minimal impact
- refactor `activity.parent_activities` by instead telling the `acts_as_tree` gem. This gem allows us find out about ancestors and children for any given node without us having to have checks on the level of each activitiy to infer how many parents might exist - quite a fragile check. It also makes it easy to extend our hierarchy by adding or removing levels in the future.
- preload activity parents when querying collections of activities (unless it's a fund as there's no current parent)

As mentioned in the commit that added [acts_as_tree](https://github.com/amerine/acts_as_tree). I investigated Ancestry and Acts_as_nested as two alternative options first. Ready the commit for more info on why I ended up picking acts_as_tree.

## Screenshots of UI changes

I created a temporary view to demonstrate what the impact of maintaining a tree structure. For 1945 activities of all levels it took the browser 2.02 seconds in total to load - this reduced to 1.3 seconds after replacing the lost parent_id database index.

The code looked like this:

```
%table.govuk-table
  %thead.govuk-table__head
    %tr.govuk-table__row
      %th.govuk-table__header
        Title
      %th.govuk-table__header
        Fund
      %th.govuk-table__header
        Programme
      %th.govuk-table__header
        Project

  %tbody.govuk-table__body
    - @activities.each do |activity|
      %tr.govuk-table__row
        %td.govuk-table__cell= activity.title
        %td.govuk-table__cell= activity.parent_activities.first&.title
        %td.govuk-table__cell= activity.parent_activities.second&.title
        %td.govuk-table__cell= activity.parent_activities.third&.title

```

I think the case can be made that a page with 100+ activities all rendered at once is not usable, and that pagination or downloading to a spreadsheet would be more appropriate. I thought it was a useful exercise to see how performant the view would be in an extreme case and I think 2 seconds is a lot better than I feared.

![Screenshot 2020-06-26 at 15 17 52](https://user-images.githubusercontent.com/912473/85867026-393b1a80-b7c0-11ea-8054-70e7457717a7.png)

Rendered the same amount of content on `develop` without these changes the time is 5.3 seconds, up from 2 seconds. 

![Screenshot 2020-06-26 at 15 22 23](https://user-images.githubusercontent.com/912473/85867558-fb8ac180-b7c0-11ea-992a-ba2a541b3ece.png)

![Screenshot_2020-06-26 — Report your official development assistance](https://user-images.githubusercontent.com/912473/85866289-12c8af80-b7bf-11ea-8be0-272f644c95fd.png)

(there were sub-projects in this page but the screenshot cropped it :) )

### 

## Next steps

- [ ] Is an ADR required? An ADR should be added if this PR introduces a change to the architecture.
- [x] Is a changelog entry required? An entry should always be made in `CHANGELOG.md`, unless this PR is a small tweak which has no impact outside the development team.
- [ ] Do any environment variables need amending or adding?
- [ ] Have any changes to the XML been checked with the IATI validator? See [XML Validation](https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/develop/doc/xml-validation.md)