@tahb 2020-07-01
@mec I've now:

1. updated the migration with `save!(validate: false)`
2. configured Bullet to only whitelist `unused_eager_loading` in the 3 service classes for `Activity.parent`. It should flag for anything else, i.e `Activity.includes(:children)`, now or in the future. I've tested that this config overwride isn't persisted to the rest of the test suite
3. I've added some more tests into each service class for the addition of `includes` for good measure
4. I noticed a second place Bullet whitelists existed and settled for merging the 2 lists together for now

Hopefully this is ready for another look!