@leeky 2020-07-02
@tahb For consistency, wouldn't it better to rename `activity_id` to `parent_activity_id`, as this name is in use in `Budget`, `Transaction` and `PlannedDisembursement`? 

Then you can drop the `Activity#parent_activity` method altogether, and just use that association as-is?