@tahb 2020-07-02
@leeky `activity_id` is being renamed to`parent_id` as it appears to be a conventional column name that all the tree gems use. It feels right to keep with the convention and `parent` and `children` makes sense to me when associating objects of the same type up and down a tree.

I did notice the `parent_activity_id` column on Budgets etc. I felt that the association between Activity->Budget is not tree shaped as Activity->Activity is, and naming the columns the same may cause confusion. Arguable Budget could be renamed to have an `activity_id`, the fact it has to also include `parent` is a plaster because of a naming clash with the PublicActivity gem.